Simplify kernel
Remove abstraction of formulas on evs and the type itself, as this
explicit encoding is unlikely to be necessary.

What we previously called a sort of (as of yet untreated in any
substantial manner) universal quantifier is to be treated using the
universal mechanisms of lambda Prolog, but becomes confused with, and
behaves like, a nabla: a fact that the constructor is renamed to
reflect.

(There is also support for existentials, although we do not do anything
with them yet.)