types: naturals and lists
Start restoring types by defining the naturals as a separate kind
instead of grouping everything in a formless term type. This will allow
lambda Prolog to type things for us from the outset and save us explicit
judgments, even as we need to provide typing judgments, notably to serve
as generators.

Naturals propagate and force the definition of a polymorphic list type
along the way.