Light generators, running times
Generate only a base term and let computation determine the rest. This
saves much useless effort, subject to a reasonable behavior of the
computational predicates.

With this, all bugs in the STLC Abella codebase based on PLT-Redex are
found. Most are very fast. Bugs #5 and #6 need to go deeper and can take
much longer. The culprit seems to be the type argument in `lam`
abstractions, which uses the leeway in term size to generate all sorts
of repetitive junk. If we comment the `lam` line in the generator, both
bugs take about 2 minutes in my machine. Interestingly, Teyjus does not
appear to be much slower than ELPI in this situation.