Add equality to interpreter and checker
Proper support of single-clause prog definitions, where the various
disjunctive cases are encoded at the level of the clause body and not as
independent clauses, requires case-analyzing pattern matching,
previously done at the head, to be pushed and made explicit as part of
the clause body, expressed by the unification-effecting equality.

Currently, this is implicitly allowed by both the interpreter and the
kernel, which does not offer the possibility to restrict this operation
by an expert predicate.

Existing "quick" FPC definitions now need to treat the or connective
that will appear explicitly in prog clauses, whereas before this was
always enforced in separate clauses for the same predicate. The
convention is hereby reversed and illustrated by is_nat.