A minimal randomized FPC and adjusted example
The absolute simplest proof of concept reads coin tosses from the
environment, assumes every disjunction in a prog is an applicable
choice point, and considers all distributions to be an even 50-50 split.
This obviates the need for probability distributions on the programs, as
well as of oracle strings. Multiple tries are not attempted.

The basic example on list ordering is reproduced with a sequence of coin
tosses that generate a counterexample.