Perform a number of random tries in testing
The randomized certificate is wrapped in another certificate that
iterates it, i.e., spawns it in sequence a given number of times,
performing said number of independent generations of randomized test
inputs.

This trick is not the cleanest in the book because the unpacking of the
certificate that wraps the randomized one in a number of tries may occur
in a number of different proof rules, so some duplication is potentially
needed. Here by assumption the possible initial rules do not overlap
with those rules where the randomized certificate has any interesting
associated logic, so the more inelegant ramifications of this approach
are avoided.

The example is modified to (continue to) work with the new top-level
certificate constructor. This continues to work seamlessly with the
external randomizer.