Entry points for automated randomized tests
First, add a second argument to the driver script to provide a
user-defined entry point for testing, which to avoid confusion should be
uniquely defined for any given execution (otherwise each of the clauses
will be run, increasing noise and confusion).

In addition to that extension to the driver script, remove some noise
reporting non-matching lines.

Second, for the CFG examples, add entry points to the randomized tests
and an adequate range of tries (in my local machine and for the given
default seed; these may vary). Moreover, because the entry points are
nullary predicates, wrap the main randomized predicate and print the
results of a successful test run, which will not be reported by lambda
Prolog otherwise. (Alternatively, modify the script to take a
parameterized argument to --exec.)