Predicate definitions as lists of named clauses
Modify the definition of clauses to better support combinators, in
particular for the generation of random data. Namely, instead of having
`prog` predicates defined as either (a) lists of `prog` clauses, or (b)
disjunctive clause bodies; encode the disjunctive nature of clause
bodies as a list of `prolog` clause bodies, which are moreover given
names -- assumed to be unique within each predicate -- to refer to them
symbolically.

Both kernel interpreters and standard, sample definitions, are adapted
to the new structure.

This organization is still not quite as modular as that originally
presented in (a), but compositionality is possibly by paying a modest
price.