Admit regular program clauses and named lists
It seems artificial at the moment to force all [prog] declarations to
adopt the mold of named lists of clause bodies since these are currently
only used for generation of random data. For this reason, we restore
[prog] to its original form and rename the multi-body declarations as
[progs].

For now, the interpreter must accept both types of clauses, but the
checker only treats generators written with [progs]. This restores the
modular composition of programs in a lightweight manner, given that we
expect generator clauses to be written and named together, in one go,
and for then to be part of the specification against which an
implementation is tested; their modular composition will not be commonly
employed at this point.