Variable seed in driver
Add a third (optional) paramenter to the driver script, such that if
'date' is given as a value the seed will be initialized to the current
Unix time as an easy method to obtain variable random runs. In the
absence of such an instruction, a fixed seed continues to be used for
purposes of reproducibility.