First fixed size certificates
Provide a variation on size certificates that enables the generation of
data of exactly a given size (dictated by the number of constructors
produced by unfoldings). Iterative deepening can be simulated fairly
efficiently by iterating these.

This is not as efficient as obtaining a list of candidates, ordering it
according to some criterion and supplying those in the desired order.
However, the current paradigm cannot very efficiently harness the FPC
framework to do that, although moving the generators to the meta level
may.