Refine iterative deepening search by depth
Avoid redundancies in search by dividing conjunctive search as follows.
To iterate generation of elements of up to depth N, find all the
elements with depth exactly 1, 2, ..., N, in sequence.

To find all the elements of depth exactly N, upon encountering a
conjunction, divide the search into three cases:

 1. Elements of depth exactly N on both branches.

 2. Elements of depth exactly N on the left and up to (N - 1) on the
    right.

 3. Elements of depth up to (N - 1) on the left and exactly N on the
    right.

This scheme seems to have a certain amount of overhead with respect to
the simple "up to N" scheme (which we are reusing), but is of course
much more robust.

Conversely, initial tests suggest that the overhead incurred by
iterative deepening of term size is much more negligible.