Shrinking FPC
A FPC modeling the shrinking process writes certificates as pairs of
witness terms (which we will attempt to shrink) and continuation
certificates. These continuations are either nested shrinkings or, when
the instantiation process is finish, placeholders that allow the
checking process to proceed by simple computation.

(Note that, because the list of witnesses can contain disparate types
for the terms, a simple polymorphic list, without heterogeneous types,
will not fit the bill.)

At the moment, we assume that existentials are always at the head of the
target formula, and that is the only place where we need to handle their
instantiation and shrinking -- this simplifies both the writing of the
shrinking FPC definition, and relieves us of the complexities that arise
when instantiation is spread all over the place in a branching formula.

The expert charged with the shrinking simply calls the user-defined
shrinker clauses, which could be arbitrarily complex. The unshrunk term
is admitted because a shrinking of a collection of witnesses must shrink
at least one of them, but all others may remain unshrunk. The current
certificate does not keep track of whether a shrinking has been
performed, instead choosing for simplicity to allow the initial
instantiation, with all original witnesses, to go through.

Because of the recursive form adopted by natural shrinkers, the
successful checking process should produce a sequence of reduced
counterexamples, gradually decreasing in size.