Default clause weight, uniform distributions
Use negation as failure to generate a default weight of 1 to complete
partial or absent probability distributions for program clauses.

If no weights are given for a given predidate, the random distribution
is generated, otherwise the gaps are filled out with the default of 1.
This is always done at the moment, but it could be:

 - Controlled by a parameter.

 - Modified so that only uniform distributions are generated with no
   distribution data is given, resulting in failure over partial
   distributions.

Whatever the case, one should not assume that a uniform distribution
will converge, and it could very well be the case that the generated
distribution will be ill-behaved.