Add top-level wrapper to substitution FPC
The envisioned use of this certificate is passive, i.e., a logic
variable standing for the certificate will get progressively
instantiated as the proof is built by another certificate paired with
this one. For this reason, however, it is necessary that the kernel know
which certificate it has to built, and to this end a top-level wrapper
that will always be present is added -- this is the usual trick, nothing
new or interesting going on.