Full shrinking workflow
Apply the full shrinking workflow to our simple example. Here, pair any
Quick-style certificate (e.g., a generous height, that something
shrinkworthy may be generated right away) with the extractor for
substitution witnesses. Next, take that information and pass it to the
shrinker.

In order to express this easily, the checking predicate is refactored
and parameterized over an input certificate; it continues to emit the
counterexamples as outputs. This checking predicate is called twice,
once to find a base counterexample (subject to backtracking), next to
try to shrink it (also subject to backtrack).

At the top level, now we return the two sets of counterexamples,
original and shrunk.