Work out qsize bug details
In particular, observe in certain instances the certificate is not fully
instantiated with a lower bound. This must be due to the fact that the
equality connective is introduced to handle multi-clause generators, but
this is transparent to the kernel, which does not get a chance to fully
instantiate lower bounds through an expert.

I am unsure at the moment why this becomes a problem at such a late
stage in the testing process.